split markdown code into modules, and move to Swarm.Text.Markdown
