+++ src/swarm-engine/Swarm/Game/Step/Util/Command.hs
@@ -59,7 +59,7 @@ import Swarm.Language.Module (Module)
 import Swarm.Language.Pipeline.QQ (tmQ)
 import Swarm.Language.Requirements.Type qualified as R
 import Swarm.Language.Syntax
-import Swarm.Language.Text.Markdown qualified as Markdown
+import Swarm.Text.Markdown qualified as Markdown
 import Swarm.Log
 import Swarm.Util (applyWhen)
 import System.Clock (TimeSpec)
