+++ src/swarm-lang/Swarm/Text/Markdown/Layout.hs
@@ -1,0 +1,188 @@
+{-# LANGUAGE OverloadedStrings #-}
+
+-- |
+-- SPDX-License-Identifier: BSD-3-Clause
+--
+-- Converting Documents to token streams + layout.
+module Swarm.Text.Markdown.Layout where
+
+import Data.Char (isSpace)
+import Data.List (intercalate)
+import Data.List.NonEmpty (NonEmpty (..))
+import Data.List.NonEmpty qualified as NE
+import Data.Text (Text)
+import Data.Text qualified as T
+import Swarm.Pretty (PrettyPrec (..), prettyText, prettyTextLine, prettyTextWidth)
+import Swarm.Text.Markdown.Document
+import Swarm.Text.Markdown.Token
+import Swarm.Util (chopNE, spanMaybe)
+
+------------------------------------------------------------
+-- Utility functions on text
+------------------------------------------------------------
+
+-- | Split text into individual whitespace and non-whitespace tokens.
+--   Each newline is made into an individual token; other whitespace
+--   are grouped into consecutive equal characters.
+--
+-- >>> tokenize "Hello   there, \n\nworld!"
+-- ["Hello","   ","there,"," ","\n","\n","world!"]
+tokenize :: Text -> [Text]
+tokenize t = case T.uncons t of
+  Nothing -> []
+  Just (c, t')
+    | c == '\n' -> "\n" : tokenize t'
+    | isSpace c -> let (spc, rest) = T.span (== c) t in spc : tokenize rest
+    | otherwise -> let (tok, rest) = T.span (not . isSpace) t in tok : tokenize rest
+
+--------------------------------------------------------------
+-- Document -> token stream conversion
+--------------------------------------------------------------
+
+-- A Document is intended to have e.g. hierarchical structure
+-- (especially once we add things like links, lists, etc.); for
+-- rendering purposes we want to first convert a structured Document
+-- into a simple token stream, split into lines at some specified
+-- maximum line width.
+
+------------------------------------------------------------
+-- ToStream class
+------------------------------------------------------------
+
+-- | Things that can be converted into a stream of tokens, possibly
+--   taking into account an optional line width.
+class ToStream a p where
+  -- | Convert to a stream of tokens, taking into account an optional line width.
+  --   a specified line width by inserting Newline tokens.  If no line
+  --   width is given, no extra Newline tokens will be inserted
+  --   (though some may still be generated by e.g. code blocks).
+  toStream :: Maybe Int -> a -> [Token' p]
+
+------------------------------------------------------------
+-- Node -> token stream conversion
+------------------------------------------------------------
+
+-- | Convert a document node into a token stream.  The line width is
+--   used only for pretty-printing code blocks; all other nodes are
+--   simply turned into a linear stream without newline tokens.
+instance PrettyPrec a => ToStream (Node a) Layout where
+  toStream :: PrettyPrec a => Maybe Int -> Node a -> [Token]
+  toStream mw = \case
+    -- Text and Raw nodes just turn into single tokens, with special
+    -- cases for Text nodes to recognize spaces and punctuation.
+    LeafText a t
+      | T.all isSpace t -> [SoftSpace]
+      | otherwise -> foldr applyAttr [TextToken t] a
+    LeafRaw a t -> applyAttr (Raw a) . map (mkToken True) . tokenize $ t
+    -- Inline code nodes get pretty-printed as a single line, then split
+    -- into code tokens separated by soft spaces.
+    -- TODO (#574): don't use tokenize, the pretty-printer should directly emit tokens!
+    LeafCode c -> applyAttr Code . map (mkToken False) . tokenize . prettyTextLine $ c
+    -- Code blocks get pretty-printed onto multiple lines using an
+    -- appropriate line width, then split into code tokens with hard
+    -- spaces.
+    LeafCodeBlock _i c -> applyAttr Code . map (mkToken True) . tokenize $ maybe (prettyText c) (prettyTextWidth c) mw
+   where
+    applyAttr attr ts = PushAttr attr : ts ++ [PopAttr]
+    mkToken hard = \case
+      "\n" -> Newline
+      t
+        | T.all isSpace t -> if hard then HardSpace (T.length t) else SoftSpace
+        | otherwise -> TextToken t
+
+------------------------------------------------------------
+-- Token stream normalisation
+------------------------------------------------------------
+
+-- | Final normalisation step on a token stream:
+--     - Get rid of special token types like HardSpace, SoftSpace, Glue, and EmptyToken
+--     - Chunk text tokens together as much as possible, to cut down on e.g. number of brick widgets generated
+normalise :: [Token] -> [OutputToken]
+normalise = mergeTokens . concatMap normaliseToken
+ where
+  -- Normalise a single token.
+  normaliseToken :: Token -> [OutputToken]
+  normaliseToken = \case
+    -- Tokens that are already normalised
+    TextToken t -> [TextToken t]
+    PushAttr a -> [PushAttr a]
+    PopAttr -> [PopAttr]
+    Newline -> [Newline]
+    Para -> [Para]
+    -- Turn spaces into text tokens.
+    SoftSpace -> [" "]
+    HardSpace n -> [TextToken (T.replicate n " ")]
+    -- Expand glue tokens.
+    Glue ts -> concatMap normaliseToken (NE.toList ts)
+    -- Filter out empty tokens.
+    EmptyToken -> []
+
+  -- Now, merge as many consecutive text tokens as we can.
+  mergeTokens :: [OutputToken] -> [OutputToken]
+  mergeTokens = chopNE nextToken
+
+  nextToken :: NonEmpty OutputToken -> (OutputToken, [OutputToken])
+  nextToken (t :| ts) = case spanMaybe getTokenText (t : ts) of
+    ([], _) -> (t, ts)
+    (txts, rest) -> (TextToken (T.concat txts), rest)
+
+------------------------------------------------------------
+-- Paragraph -> token stream conversion + layout
+------------------------------------------------------------
+
+-- | Convert a paragraph into a token stream.  The line width is used
+--   to flow the paragraph by inserting newline tokens appropriately.
+--   The resulting token stream is also normalized by merging together
+--   consecutive tokens of the same type as much as possible, e.g. to cut
+--   down on the number of Brick widgets to be generated.
+instance PrettyPrec a => ToStream (Paragraph a) Output where
+  toStream mw =
+    normalise
+      . maybe id (\w -> splitter w w) mw
+      . glueTokens
+      . concatMap (toStream mw)
+      . nodes
+   where
+    -- Split a token stream into lines by inserting newline tokens,
+    -- and ensure there are no SoftSpace tokens at the beginning or
+    -- end of any line.
+    --
+    -- The first Int is the max line width for each line.
+    -- The second Int tells us how much width remains on the current line.
+    splitter :: Int -> Int -> [Token] -> [Token]
+    splitter width remaining = \case
+      [] -> []
+      -- If we encounter an existing newline token, just emit it and
+      -- move on to the next line, resetting the available width
+      Newline : ts -> Newline : splitter width width ts
+      -- Special handling for SoftSpace
+      SoftSpace : ts -> case ts of
+        -- Discard a SoftSpace at the end of a line
+        [] -> []
+        Newline : ts' -> Newline : splitter width width ts'
+        (t : ts')
+          -- Discard a SoftSpace at the beginning of a line
+          | width == remaining -> splitter width remaining ts
+          -- If we can emit the space + next token, do so
+          | 1 + tokenWidth t <= remaining -> SoftSpace : t : splitter width (remaining - 1 - tokenWidth t) ts'
+          -- Otherwise, discard the SoftSpace + move to the next line.
+          | otherwise -> Newline : splitter width width ts
+      -- In the general case, check if we can emit the next token
+      t : ts
+        | tokenWidth t <= remaining -> t : splitter width (remaining - tokenWidth t) ts
+        -- ...but if the next token doesn't fit and we are at the
+        -- beginning of the line, we must chop it into pieces to force
+        -- progress
+        | width == remaining ->
+            let (t1, t2) = splitTokenAt width t in t1 : Newline : splitter width width (t2 : ts)
+      -- Finally, if nothing fits, emit a Newline and proceed to the next line
+      ts -> Newline : splitter width width ts
+
+------------------------------------------------------------
+-- Top-level Document -> token stream conversion
+------------------------------------------------------------
+
+-- | Convert an entire document into a token stream, inserting
+--   paragraph breaks in between consecutive paragraphs.
+instance PrettyPrec a => ToStream (Document a) Output where
+  toStream mw = intercalate [Para] . map (toStream mw) . paragraphs
