+++ src/swarm-scenario/Swarm/Game/Scenario.hs
@@ -107,7 +107,7 @@ import Swarm.Language.JSON (withModuleProvenance)
 import Swarm.Language.Module (Module, ModuleCtx, ModuleImports)
 import Swarm.Language.Pipeline (Processable (..), processSyntax)
 import Swarm.Language.Syntax (Anchor, ImportPhaseFor, Phase (..), SwarmType, Syntax)
-import Swarm.Language.Text.Markdown (Document)
+import Swarm.Text.Markdown (Document)
 import Swarm.Pretty (prettyText)
 import Swarm.ResourceLoading (getDataFileNameThrow)
 import Swarm.Util (binTuples, commaList, failT, quote)
