+++ src/swarm-scenario/Swarm/Game/Scenario/Topography/WorldDescription.hs
@@ -46,7 +46,7 @@ import Swarm.Game.Universe (SubworldName (DefaultRootSubworld))
 import Swarm.Game.World.DSL
 import Swarm.Language.Pipeline (Processable (..))
 import Swarm.Language.Syntax (Phase (Raw), Syntax)
-import Swarm.Language.Text.Markdown (Document, fromText)
+import Swarm.Text.Markdown (Document, fromText)
 import Swarm.Pretty (prettyString)
 import Swarm.Util.Erasable (Erasable (..))
 import Swarm.Util.Yaml
