+++ src/swarm-tui/Swarm/TUI/Model.hs
@@ -126,7 +126,7 @@ import Swarm.Game.State.Substate
 import Swarm.Game.Tick (TickNumber (..))
 import Swarm.Game.World (Seed)
 import Swarm.Language.Syntax (Phase (..))
-import Swarm.Language.Text.Markdown qualified as Markdown
+import Swarm.Text.Markdown qualified as Markdown
 import Swarm.Log
 import Swarm.TUI.Inventory.Sorting
 import Swarm.TUI.Model.DebugOption (DebugOption)
